+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### PulseUI
 
+- Add public `StoreDetailsView`
 - Fix an issue where search in text viewer would occasionally stop highlighting the results
 - Add a missing public `InsightsView` initializer 
 - Add cookies section and details screens to the request summary - [#124](https://github.com/kean/Pulse/pull/124)